Make class be treated differently than other attributes
wf_tags:emit_tag previously treated each attribute kind of like how the
class attribute should only be treated: a list of atoms becomes a list
of space-delimited atoms (with the list being reversed as well).

This would lead to issues where attribute values specified with iolists
would be split up weirdly, and reversed.  This fixes that, and makes
attributes better to work with due to being more supportive of iolists