Don't lose actions during a comet/ws changeover
There was a bug where a comet request started before the websocket
connection was accepted, that some actions would be "lost" in the
Accumulator (since the websocket functionality doesn't rely on the
Accumulator to store actions).  This ensures that when the websocket
request starts, that 1) it immediately picks up any other requests, and
2) any requests that somehow manage to make it into the Accumulator will
get picked up during a flush.