Add isheader-attribute to tablecell-element
In order to generate html-tables sometimes we want
to follow the thead/tr/th structure.

In the current version of nitrogen one cannot choose
between th- or td-elements. This patchs tries to fix this.

Alternatives tried: copy the tablecell into a tablehcell-module.
The difference being that tablehcell would be used for headers.

The code felt messy and since these changes are so minor this
became the first attempt.

Link:
 - https://www.w3schools.com/tags/tag_thead.asp