Use Unicode-aware to-binary conversion when input might require it.
In certain situations we've been passing an input to `wf:to_binary()`
that may contain text with code points that form multi-byte sequences
when encoded.

For example, the string `[127344, 127345, 127346]` corresponds to the
string input "🅰🅱🅲", but corresponds to the UTF-8 byte sequence
`[240,159,133,176,240,159,133,177,240,159,133,178]`.

Problematically, while we can easily convert from binary to a list in
Erlang with `binary_to_list/1`, we cannot so easily convert from list
to binary with `list_to_binary/1`, because that function operates on
the `iolist()` datatype, which operates on bytes, not on abstract
numbers.

Per the documentation for `iolist()` [1],

> In most use cases you want to use iodata() instead of this type.

Per the documentation for `iodata()` [2],

> To transcode a string() or unicode:chardata() to iodata()
> you can use unicode:characters_to_binary/1.

And the reason is that `unicode:characters_to_binary/1` will properly
convert the code points in a string-list into the proper UTF-8 byte
sequence that's necessary for the final output.

In this patch we're calling `wf:to_unicode_binary/1` in places where
I found us to be passing likely-Unicode data. There might be more
places that I didn't find, and one might argue that in some of the
cases I changed, we don't expect Unicode data. The risk of the former
is that this patch isn't a full fix for every case, but that doesn't
imply these fixes are invalid; the risk of the latter is that we could
choke on non-Unicode data, but that should only be the case _if_ the
unpatched code would have also failed.

Noted in the erlang forums [3]

[1]: https://www.erlang.org/doc/man/erlang.html#type-iolist
[2]: https://www.erlang.org/doc/man/erlang.html#type-iodata
[3]: https://erlangforums.com/t/problem-with-non-latin-symbols-in-nitrogen-element-names/2087