Notice Bar: Remove dependency on jQuery
Rework the notice bar so that it relies on DOM operations and CSS
transitions for animations instead of jQuery. When hiding the
notice bar this leaves the Element in the DOM so that if and when
it re-opens it will animate instead of appearing suddenly, which
is the current behavior.

In order to support the refactor the class naming scheme has been
changed so that the notice bar is identified by its `id` attribute
and the classes only desigate whether the notice bar is visible
and what type of notice it is: e.g. `is-error` or `is-ok`.